fix depricated apis and enhanced bash scripts

+++ build_vsphere.sh
@@ -3,10 +3,17 @@
 set -e # fail fast
 set -x # print commands
 
-git clone https://github.com/laidbackware/vsphere-lab-deploy.git
+# Check if running as sudo
+if [[ $EUID > 0 ]]; then
+  echo "Please run as sudo/root"
+  exit 1
+fi
+
+date
+
+git clone https://github.com/laidbackware/vsphere-lab-deploy.git --branch=optimize_build
 
 cd vsphere-lab-deploy
-git checkout optimize_build
 
 rm ./*yml
 
@@ -14,4 +21,4 @@ cp ../vsphere-answerfile.yml ./answerfile.yml
 cp ../vsphere-deploy.yml .
 ansible-playbook vsphere-deploy.yml
 
-cd ..
+date
